Rename RequestTracer to Tracer
