+++ docs/trace/api/context_tracer.rst
@@ -1,6 +1,6 @@
 Context Tracer
 ==============
 
-.. automodule:: opencensus.trace.tracer.context_tracer
+.. automodule:: opencensus.trace.tracers.context_tracer
   :members:
   :show-inheritance:
