+++ tests/unit/trace/tracers/test_base_tracer.py
@@ -14,7 +14,7 @@
 
 import unittest
 
-from opencensus.trace.tracer import base
+from opencensus.trace.tracers import base
 
 
 class TestBaseTracer(unittest.TestCase):
